+++ packages/client/src/app/App.tsx
@@ -6,7 +6,7 @@ import { Main } from '../pages/main'
 import { SignIn } from '../pages/sign-in'
 import { SignUp } from '../pages/sign-up'
 import { Forum } from '../pages/forum'
-import { LeaderBoard } from '../pages/main/modules/leader-board'
+import { LeaderBoard } from '../pages/main/modules/leaderBoard'
 
 function RequireAuth({ children }: { children: JSX.Element }) {
   // Когда будем подключать аунтификацию - будет использоваться обертка, для проверки авторизован ли пользователь,
