+++ packages/client/src/pages/main/modules/leaderBoard/LeaderBoard.tsx
@@ -1,20 +1,26 @@
-import React from 'react'
+import React, { useState, useEffect } from 'react'
 import styles from './leaderBoard.module.scss'
 import Container from 'react-bootstrap/Container'
-import Card from 'react-bootstrap/Card'
 import DataTable, { TableColumn } from 'react-data-table-component'
 import dummyData from './dummy.json'
+import { TopPlayerCard } from './components/TopPlayerCard'
+import { LeaderBoardAvatar } from './components/LeaderBoardAvatar'
 
-interface DataRow {
+export interface IDataRow {
   rank: number;
   avatar: string;
   user: string;
   score: number;
 }
 
+export interface IData extends IDataRow {
+  id: number;
+}
+
 const LeaderBoard: React.FC = () => {
-  const defaultAvatar = 'https://cdn-icons-png.flaticon.com/512/149/149071.png';
-  const columns: TableColumn<DataRow>[] = [
+  const [topPlayersData, setTopPlayersData] = useState<IData[]>([]);
+  const data = dummyData;
+  const columns: TableColumn<IDataRow>[] = [
     {
       name: 'Rank',
       selector: row => row.rank,
@@ -22,7 +28,7 @@ const LeaderBoard: React.FC = () => {
     },
     {
       name: 'Avatar',
-      cell: (row) => <img className={styles.table_img} src={row.avatar || defaultAvatar} />,
+      cell: (row) => <LeaderBoardAvatar {...row}/>,
       sortable: false,
     },
     {
@@ -37,23 +43,17 @@ const LeaderBoard: React.FC = () => {
     },
   ];
 
-  const data = dummyData;
-  const topPlayersData = dummyData.sort((a, b) => b.score - a.score).slice(0, 3);
+  useEffect(() => {
+    const topPlayers = dummyData.sort((a, b) => b.score - a.score).slice(0, 3);
+    setTopPlayersData(topPlayers);
+  }, [dummyData])
 
   return (
     <Container className={`p-5 d-flex flex-column ${styles.container}`}>
       <div className={styles.wrapper_top}>
         <h1 className='mb-3'>Top Players</h1>
         <div className='d-flex justify-content-evenly mb-3'>
-          {topPlayersData && topPlayersData.map(data => 
-            <Card className={styles.card} key={data.id}>
-              <Card.Img className={styles.card_img} src={data.avatar} />
-              <Card.Body>
-                <Card.Title>{data.user}</Card.Title>
-                <Card.Text>{data.score}</Card.Text>
-              </Card.Body>
-            </Card>
-          )} 
+          {topPlayersData && topPlayersData.map(data => <TopPlayerCard key={data.id} {...data}/>)} 
         </div>
       </div>
       <div className={styles.wrapper_bottom}>
