+++ packages/client/src/pages/main/modules/leaderBoard/components/LeaderBoardAvatar/index.ts
@@ -1,0 +1,1 @@
+export { default as LeaderBoardAvatar } from './LeaderBoardAvatar'
