+++ packages/client/src/pages/main/modules/leaderBoard/components/LeaderBoardAvatar/leaderBoardAvatar.module.scss
@@ -1,0 +1,5 @@
+.table_img {
+  width: 40px;
+  height: 40px;
+  border-radius: 50%;
+}
