+++ packages/client/src/pages/main/modules/leaderBoard/components/TopPlayerCard/index.ts
@@ -1,0 +1,1 @@
+export { default as TopPlayerCard } from './TopPlayerCard'
