+++ packages/client/src/pages/main/modules/leaderBoard/components/TopPlayerCard/topPlayerCard.module.scss
@@ -1,19 +1,3 @@
-.container {
-  height: 100vh;
-  width: 70vw;
-
-  .wrapper_top {
-    height: calc(50% - 1.5rem);
-  }
-
-  .wrapper_bottom {
-    height: calc(50% - 1.5rem);
-    width: 75%;
-    margin-left: auto;
-    margin-right: auto;
-  }
-}
-
 .card {
   width: 20%;
   box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(0, 0, 0, 0.19);
@@ -28,10 +12,4 @@
     border-radius: 50%;
     margin-top: 10px;
   }
-}
-
-.table_img {
-  width: 40px;
-  height: 40px;
-  border-radius: 50%;
 }
