+++ packages/client/src/pages/main/modules/leaderBoard/leaderBoard.module.scss
@@ -1,0 +1,15 @@
+.container {
+  height: 100vh;
+  width: 70vw;
+
+  .wrapper_top {
+    height: calc(50% - 1.5rem);
+  }
+
+  .wrapper_bottom {
+    height: calc(50% - 1.5rem);
+    width: 75%;
+    margin-left: auto;
+    margin-right: auto;
+  }
+}
