+++ packages/client/src/utils/const/const.ts
@@ -1,0 +1,1 @@
+export const AVATAR_PLACEHOLDER = 'https://cdn-icons-png.flaticon.com/512/149/149071.png'
